requirements.txt
asgiref==3.4.1
autopep8==1.5.7
Django==3.2.12
django-crispy-forms==1.13.0
django-debug-toolbar==3.2.2
django-email-users==1.1.0
django-grappelli==3.0.6
django-summernote==0.8.11.6
Pillow==9.5.0
psycopg2==2.9.6
pycodestyle==2.8.0
pytz==2021.3
signals==0.0.2
sqlparse==0.4.2
toml==0.10.2
